actions/average_payment_time.go
package actions

import (
	"database/sql"
	"net/http"

	"github.com/Iledant/iris-propera/models"
	"github.com/kataras/iris"
)

// GetAvgPmtTimes handles the get request to fetch the average payment time of
// the past 12 monthes.
func GetAvgPmtTimes(ctx iris.Context) {
	var resp models.AvgPmtTimes
	db := ctx.Values().Get("db").(*sql.DB)
	if err := resp.GetAll(db); err != nil {
		ctx.StatusCode(http.StatusInternalServerError)
		ctx.JSON(jsonError{"Durée moyenne de paiement, requête : " + err.Error()})
		return
	}
	ctx.StatusCode(http.StatusOK)
	ctx.JSON(resp)
}
